% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/energy_lattice.R
\name{energy_lattice}
\alias{energy_lattice}
\title{Energy across a binary lattice}
\usage{
energy_lattice(lattice)
}
\arguments{
\item{lattice}{lattice for which energy is required.}
}
\description{
Calculate the energy across a binary lattice of 0s and 1s.
}
