% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genbinom.R, R/get_neighbours.R
\name{GenBinomial}
\alias{GenBinomial}
\alias{dgenbinom}
\alias{pgenbinom}
\alias{qgenbinom}
\alias{rgenbinom}
\alias{get_neighbours}
\title{The Generalised Binomial Distribution}
\usage{
dgenbinom(g, pi, theta, size)

pgenbinom(g, pi, theta, size, lower.tail = TRUE)

qgenbinom(p, pi, theta, size, lower.tail = TRUE)

rgenbinom(n, pi, theta, size)

get_neighbours(x, y, size)
}
\arguments{
\item{g}{vector of quantiles.}

\item{pi}{probability of success of the bernoulli.}

\item{theta}{probability of success on each binomial trial.}

\item{size}{vector of dimensions of the lattice in the form \code{c(x,y)}.}

\item{lower.tail}{logical; if TRUE (default), probabilities are
\eqn{Pr[X \leq x]}{Pr[X \le x]}, otherwise \eqn{Pr[X > x]}{Pr[X > x]}.}

\item{p}{vector of probabilities.}

\item{n}{number of observations.}

\item{x}{x coordinate of node.}

\item{y}{y coordinate of node.}
}
\description{
Density, distribution function, quantile function and random
    generation for the generalised binomial distribution with parameters
    \code{pi}, \code{theta} and \code{size}.

Density, distribution function, quantile function and random
    generation for the generalised binomial distribution with parameters
    \code{pi}, \code{theta} and \code{size}.
}
\details{
The generalised binomial distribution with \code{size = n},
    \code{pi = pi} and \code{theta = theta} has density
    \deqn{Pr(G = g) = \begin{cases} x, g = 0 \\ y, g > 0\end{cases}}

For a given node on a lattice of dimensions specified by \code{size}
    calculate the neighbouring nodes according to the von Neumann method.
}
