% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genbinomquantities.R
\name{GenBinomialQuantities}
\alias{GenBinomialQuantities}
\alias{genbinom_expectation}
\alias{genbinom_variance}
\title{The Generalised Binomial Quantities}
\usage{
genbinom_expectation(pi, theta, size)
}
\arguments{
\item{pi}{probability of success of the bernoulli.}

\item{theta}{probability of success on each binomial trial.}

\item{size}{number of trials.}
}
\description{
Expectation and variance for the generalised binomial
    distribution with parameters \code{pi}, \code{theta} and \code{size}.
}
\details{
This function derives the expectation and variance for
    the generalised binomial distribution with \code{size = n},
    \code{pi = pi} and \code{theta = theta}
}
